refactor: session 에서 저장된 id return 하도록 작업

+++ src/main/java/nextstep/courses/domain/SessionRepository.java
@@ -3,7 +3,7 @@ package nextstep.courses.domain;
 import java.util.List;
 
 public interface SessionRepository {
-    int save(Session session);
+    long save(Session session);
 
     Session findById(Long id);
